+++ etna/transforms/feature_selection/gale_shapley.py
@@ -236,6 +236,7 @@ class GaleShapleyFeatureSelectionTransform(BaseFeatureSelectionTransform):
         top_k: int,
         features_to_use: Union[List[str], Literal["all"]] = "all",
         use_rank: bool = False,
+        return_features: bool = False,
         **relevance_params,
     ):
         """Init GaleShapleyFeatureSelectionTransform.
@@ -251,8 +252,10 @@ class GaleShapleyFeatureSelectionTransform(BaseFeatureSelectionTransform):
             if "all" value is given, all columns are used
         use_rank:
             if True, use rank in relevance table computation
+        return_features:
+            bool, indicates whether to return features or not.
         """
-        super().__init__(features_to_use=features_to_use)
+        super().__init__(features_to_use=features_to_use, return_features=return_features)
         self.relevance_table = relevance_table
         self.top_k = top_k
         self.use_rank = use_rank